ENH: move exceptions closer to "home"
